requirements.txt
amqp==2.6.1
aniso8601==7.0.0
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.7
Django==3.1.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
factory-boy==2.12.0
Faker==4.1.2
graphene==2.1.8
graphene-django==2.13.0
graphql-core==2.3.2
graphql-relay==2.0.1
kombu==4.6.11
promise==2.3
psycopg2-binary==2.8.5
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
Unidecode==1.1.1
vine==1.3.0
